+++ src/EnergyPlus/ChillerGasAbsorption.cc
@@ -132,14 +132,14 @@ void GasAbsorberSpecs::simulate(
     DataPlant::BrLoopType brIdentity(DataPlant::BrLoopType::NoMatch);
 
     int branchTotalComp = state.dataPlnt->PlantLoop(calledFromLocation.loopNum)
-                              .LoopSide[static_cast<int>(calledFromLocation.loopSideNum)]
+                              .LoopSide[calledFromLocation.loopSideNum]
                               .Branch(calledFromLocation.branchNum)
                               .TotalComponents;
 
     for (int iComp = 1; iComp <= branchTotalComp; iComp++) {
         // kind of a hacky way to find the location of this, but it's what plantloopequip was doing
         int compInletNodeNum = state.dataPlnt->PlantLoop(calledFromLocation.loopNum)
-                                   .LoopSide[static_cast<int>(calledFromLocation.loopSideNum)]
+                                   .LoopSide[calledFromLocation.loopSideNum]
                                    .Branch(calledFromLocation.branchNum)
                                    .Comp(iComp)
                                    .NodeNumIn;
@@ -200,14 +200,14 @@ void GasAbsorberSpecs::getDesignCapacities(
     bool matchfound(false);
 
     int branchTotalComp = state.dataPlnt->PlantLoop(calledFromLocation.loopNum)
-                              .LoopSide[static_cast<int>(calledFromLocation.loopSideNum)]
+                              .LoopSide[calledFromLocation.loopSideNum]
                               .Branch(calledFromLocation.branchNum)
                               .TotalComponents;
 
     for (int iComp = 1; iComp <= branchTotalComp; iComp++) {
         // kind of a hacky way to find the location of this, but it's what plantloopequip was doing
         int compInletNodeNum = state.dataPlnt->PlantLoop(calledFromLocation.loopNum)
-                                   .LoopSide[static_cast<int>(calledFromLocation.loopSideNum)]
+                                   .LoopSide[calledFromLocation.loopSideNum]
                                    .Branch(calledFromLocation.branchNum)
                                    .Comp(iComp)
                                    .NodeNumIn;
@@ -255,7 +255,7 @@ void GasAbsorberSpecs::onInitLoopEquip(EnergyPlusData &state, const PlantLocatio
 
     // kind of a hacky way to find the location of this, but it's what plantloopequip was doing
     int BranchInletNodeNum =
-        state.dataPlnt->PlantLoop(calledFromLocation.loopNum).LoopSide[static_cast<int>(calledFromLocation.loopSideNum)].Branch(calledFromLocation.branchNum).NodeNumIn;
+        state.dataPlnt->PlantLoop(calledFromLocation.loopNum).LoopSide[calledFromLocation.loopSideNum].Branch(calledFromLocation.branchNum).NodeNumIn;
 
     if (BranchInletNodeNum == this->ChillReturnNodeNum) {       // Operate as chiller
         this->size(state);                                      // only call from chilled water loop
@@ -1690,7 +1690,7 @@ void GasAbsorberSpecs::calculateChiller(EnergyPlusData &state, Real64 &MyLoad)
         LoopNum = this->CWLoopNum;
         LoopSideNum = this->CWLoopSideNum;
         {
-            auto const SELECT_CASE_var(state.dataPlnt->PlantLoop(LoopNum).LoopSide[static_cast<int>(LoopSideNum)].FlowLock);
+            auto const SELECT_CASE_var(state.dataPlnt->PlantLoop(LoopNum).LoopSide[LoopSideNum].FlowLock);
             if (SELECT_CASE_var == DataPlant::FlowLock::Unlocked) { // mass flow rates may be changed by loop components
                 this->PossibleSubcooling = false;
                 lCoolingLoad = std::abs(MyLoad);
@@ -1980,7 +1980,7 @@ void GasAbsorberSpecs::calculateHeater(EnergyPlusData &state, Real64 &MyLoad, bo
         //    cooling load taken by the chiller, and
         //    supply temperature
         {
-            auto const SELECT_CASE_var(state.dataPlnt->PlantLoop(LoopNum).LoopSide[static_cast<int>(LoopSideNum)].FlowLock);
+            auto const SELECT_CASE_var(state.dataPlnt->PlantLoop(LoopNum).LoopSide[LoopSideNum].FlowLock);
             if (SELECT_CASE_var == DataPlant::FlowLock::Unlocked) { // mass flow rates may be changed by loop components
                 lHeatingLoad = std::abs(MyLoad);
                 if (HeatDeltaTemp != 0) {
