+++ src/EnergyPlus/HVACSizingSimulationManager.cc
@@ -114,7 +114,7 @@ void HVACSizingSimulationManager::CreateNewCoincidentPlantAnalysisObject(EnergyP
 
             plantCoincAnalyObjs.emplace_back(PlantLoopName,
                                              i,
-                                             state.dataPlnt->PlantLoop(i).LoopSide[static_cast<int>(DataPlant::LoopSideLocation::Supply)].NodeNumIn,
+                                             state.dataPlnt->PlantLoop(i).LoopSide[DataPlant::LoopSideLocation::Supply].NodeNumIn,
                                              density,
                                              cp,
                                              state.dataSize->PlantSizData(PlantSizingIndex).NumTimeStepsInAvg,
