+++ src/EnergyPlus/PlantCentralGSHP.cc
@@ -1612,13 +1612,13 @@ void WrapperSpecs::initialize(EnergyPlusData &state,
             // Reset flow priority
             if (LoopNum == this->CWLoopNum) {
                 state.dataPlnt->PlantLoop(this->CWLoopNum)
-                    .LoopSide[static_cast<int>(this->CWLoopSideNum)]
+                    .LoopSide[this->CWLoopSideNum]
                     .Branch(this->CWBranchNum)
                     .Comp(this->CWCompNum)
                     .FlowPriority = DataPlant::LoopFlowStatus::NeedyIfLoopOn;
             } else if (LoopNum == this->HWLoopNum) {
                 state.dataPlnt->PlantLoop(this->HWLoopNum)
-                    .LoopSide[static_cast<int>(this->HWLoopSideNum)]
+                    .LoopSide[this->HWLoopSideNum]
                     .Branch(this->HWBranchNum)
                     .Comp(this->HWCompNum)
                     .FlowPriority = DataPlant::LoopFlowStatus::NeedyIfLoopOn;
@@ -2865,7 +2865,7 @@ void WrapperSpecs::CalcWrapperModel(EnergyPlusData &state, Real64 &MyLoad, int c
         CurCoolingLoad = std::abs(MyLoad);
         this->WrapperCoolingLoad = CurCoolingLoad;
         // Set actual mass flow rate at the nodes when it's locked
-        if (state.dataPlnt->PlantLoop(LoopNum).LoopSide[static_cast<int>(LoopSideNum)].FlowLock == DataPlant::FlowLock::Locked) {
+        if (state.dataPlnt->PlantLoop(LoopNum).LoopSide[LoopSideNum].FlowLock == DataPlant::FlowLock::Locked) {
             CHWInletMassFlowRate = state.dataLoopNodes->Node(this->CHWInletNodeNum).MassFlowRate;
         }
         if (CHWInletMassFlowRate == 0.0) GLHEInletMassFlowRate = 0.0;
@@ -2879,7 +2879,7 @@ void WrapperSpecs::CalcWrapperModel(EnergyPlusData &state, Real64 &MyLoad, int c
         CurHeatingLoad = MyLoad;
         this->WrapperHeatingLoad = CurHeatingLoad;
         // Set actual mass flow rate at the nodes when it's locked
-        if (state.dataPlnt->PlantLoop(LoopNum).LoopSide[static_cast<int>(LoopSideNum)].FlowLock == DataPlant::FlowLock::Locked) {
+        if (state.dataPlnt->PlantLoop(LoopNum).LoopSide[LoopSideNum].FlowLock == DataPlant::FlowLock::Locked) {
             HWInletMassFlowRate = state.dataLoopNodes->Node(this->HWInletNodeNum).MassFlowRate;
         }
         if (HWInletMassFlowRate == 0.0) GLHEInletMassFlowRate = 0.0;
